Separate out plugin.js to avoid circular import
