+++ packages/eslint-plugin/lib/config/polaris.js
@@ -1,4 +1,4 @@
-const shopifyPlugin = require('../..');
+const shopifyPlugin = require('../../plugin');
 
 module.exports = [
   {
